vscode_project/requirements.txt
fastapi
uvicorn[standard]
pydantic
scikit-learn
numpy
joblib
tldextract
python-whois
beautifulsoup4
textblob
textstat
spacy